+++ go/integration/cli_srv_ext_pyintegration/main.go
@@ -41,13 +41,9 @@ func realMain() int {
 	clientArgs := []string{"--port", integration.ServerPortReplace,
 		integration.SrcIAReplace, integration.DstIAReplace}
 	serverArgs := []string{"--run_server", integration.DstIAReplace}
-	var in integration.Integration
+	in := integration.NewBinaryIntegration(name, cmd, clientArgs, serverArgs, integration.StdLog)
 	if *integration.Container != "" {
-		in = integration.NewDockerIntegration(name, *integration.Container, cmd, clientArgs,
-			serverArgs, integration.StdLog)
-	} else {
-		in = integration.NewBinaryIntegration(name, cmd, clientArgs, serverArgs,
-			integration.StdLog)
+		in = integration.NewDockerIntegration(*integration.Container, in)
 	}
 	if err := integration.RunBinaryTests(in, integration.IAPairs()); err != nil {
 		fmt.Fprintf(os.Stderr, "Failed to run tests: %s\n", err)
