+++ go/lib/integration/common.go
@@ -16,39 +16,13 @@ package integration
 
 import (
 	"bufio"
-	"context"
 	"fmt"
 	"io"
-	"os"
-	"os/exec"
 	"strings"
 
 	"github.com/scionproto/scion/go/lib/addr"
 	"github.com/scionproto/scion/go/lib/log"
 	"github.com/scionproto/scion/go/lib/log/logparse"
-)
-
-const (
-	// ServerPortReplace is a placeholder for the server port in the arguments.
-	ServerPortReplace = "<ServerPort>"
-	// SrcIAReplace is a placeholder for the source IA in the arguments.
-	SrcIAReplace = "<SRCIA>"
-	// DstIAReplace is a placeholder for the destination IA in the arguments.
-	DstIAReplace = "<DSTIA>"
-	// ReadySignal should be written to Stdout by the server once it is read to accept clients.
-	// The message should always be `Listening ia=<IA>`
-	// where <IA> is the IA the server is listening on.
-	ReadySignal = "Listening ia="
-	// GoIntegrationEnv is an environment variable that is set for the binary under test.
-	// It can be used to guard certain statements, like printing the ReadySignal,
-	// in a program under test.
-	GoIntegrationEnv = "SCION_GO_INTEGRATION"
-	// portString is the string a server prints to specify the port it's listening on.
-	portString = "Port="
-)
-
-var (
-	serverPorts = make(map[addr.IA]string)
 )
 
 type LogRedirect func(name, pName string, local addr.IA, ep io.ReadCloser)
@@ -82,83 +56,4 @@ func replacePattern(pattern string, replacement string, args []string) []string 
 		}
 	}
 	return argsCopy
-}
-
-var _ Waiter = (*waiter)(nil)
-
-type waiter struct {
-	*exec.Cmd
-}
-
-func startServer(ctx context.Context, cmd string, args []string, dst addr.IA,
-	logRedirect LogRedirect) (Waiter, error) {
-
-	startCtx, cancelF := context.WithTimeout(ctx, StartServerTimeout)
-	defer cancelF()
-	r := &waiter{
-		exec.CommandContext(ctx, cmd, args...),
-	}
-	r.Env = os.Environ()
-	r.Env = append(r.Env, fmt.Sprintf("%s=1", GoIntegrationEnv))
-	ep, err := r.StderrPipe()
-	if err != nil {
-		return nil, err
-	}
-	sp, err := r.StdoutPipe()
-	if err != nil {
-		return nil, err
-	}
-	ready := make(chan struct{})
-	// parse until we have the ready signal.
-	// and then discard the output until the end (required by StdoutPipe).
-	go func() {
-		defer log.LogPanicAndExit()
-		defer sp.Close()
-		signal := fmt.Sprintf("%s%s", ReadySignal, dst)
-		init := true
-		scanner := bufio.NewScanner(sp)
-		for scanner.Scan() {
-			line := scanner.Text()
-			if strings.HasPrefix(line, portString) {
-				serverPorts[dst] = strings.TrimPrefix(line, portString)
-			}
-			if init && signal == line {
-				close(ready)
-				init = false
-			}
-		}
-	}()
-	go func() {
-		defer log.LogPanicAndExit()
-		logRedirect("Server", "ServerErr", dst, ep)
-	}()
-	err = r.Start()
-	if err != nil {
-		return nil, err
-	}
-	select {
-	case <-ready:
-		return r, err
-	case <-startCtx.Done():
-		return nil, startCtx.Err()
-	}
-}
-
-func startClient(ctx context.Context, cmd string, args []string, src addr.IA,
-	logRedirect LogRedirect) (Waiter, error) {
-
-	r := &waiter{
-		exec.CommandContext(ctx, cmd, args...),
-	}
-	r.Env = os.Environ()
-	r.Env = append(r.Env, fmt.Sprintf("%s=1", GoIntegrationEnv))
-	ep, err := r.StderrPipe()
-	if err != nil {
-		return nil, err
-	}
-	go func() {
-		defer log.LogPanicAndExit()
-		logRedirect("Client", "ClientErr", src, ep)
-	}()
-	return r, r.Start()
 }
